fix compile error in lru_cache_map test

+++ src/base/tests/lru_cache_map.cc
@@ -49,7 +49,7 @@ TEST(lru_cache_map, no_remove_condition)
   c["e3"] = 101;
   c["e4"] = 102;
 
-  EXPECT_EQ(4, c.get_size());
+  EXPECT_EQ(static_cast<size_t>(4), c.get_size());
 
   EXPECT_EQ(string("e4,e3,e2,e1"), newest(c)) << "init, newest";
   EXPECT_EQ(string("e1,e2,e3,e4"), oldest(c)) << "init, oldest";
@@ -114,7 +114,7 @@ TEST(lru_cache_map, remove_if_over_100)
   c["e3"] = 101;
   c["e4"] = 102;
 
-  EXPECT_EQ(4, c.get_size());
+  EXPECT_EQ(static_cast<size_t>(4), c.get_size());
 
   EXPECT_EQ(string("e4,e3,e2,e1"), newest(c)) << "init, newest";
   EXPECT_EQ(string("e1,e2,e3,e4"), oldest(c)) << "init, oldest";
